Move all debug menus to debug/view
