+++ game/debug/view/list_picker.lua
@@ -1,0 +1,15 @@
+
+return function(name, list, value)
+
+  return "Choose a " .. name, function(self)
+    imgui.Text("Options:")
+    imgui.PushItemWidth(160)
+    local changed, newvalue = imgui.ListBox("", value(), list, #list)
+    if changed then
+      value(newvalue)
+    end
+    imgui.PopItemWidth()
+  end
+
+end
+
